requirements.txt
uvicorn
fastapi
pydantic
pyjwt
sqlalchemy